Convert binomial (species) level occurrences to trinomial based on distributions.

Convert binomial (species) level occurrences to trinomial based on distributions.

bi_to_tri(
  species,
  presences,
  distrib_files = NULL,
  use_mcp = FALSE,
  mcp_files = NULL,
  use_clust = FALSE,
  pres_x = "long",
  pres_y = "lat",
  pres_crs = 4326,
  use_crs = NULL,
  buf = 0,
  clust_dist = 50000,
  overrides = tibble::tribble(~subspecies, ~poly, "Stipiturus malachurus intermedius",
    c("dist", "mcp"))
)

Arguments

species

Character. Name of the species (binomial) to update occurrences to trinomial for.

presences

Data frame of all binomial and trinomial presences relevant to the species. Must contain a 'subspecies' column and x, y coordinate columns corresponding to the names in pres_x & pres_y. For binomial presences 'subspecies' should be NA.

distrib_files

Data frame of relevant distribution file paths for all the subspecies (e.g. for distributions sourced from redlist, epbc etc). If no relevant distribution, then use NA. Currently, only geoparquet files are accepted. Use sfarrow::st_write_parquet to write sf objects to parquet.

use_mcp

Logical. Use a minimum convex polygon (MCP) around presences to determine relevant trinomials?

mcp_files

If use_mcp == TRUE, optional data frame of file paths of existing mcp files to be used. If use_mcp == TRUE and mcp_file == NULL, a new mcp will be constructed from presences. Currently, only geoparquet files are accepted. Use sfarrow to convert sf objects to geoparquet.

use_clust

Logical. Use clusters around around presences to determine relevant trinomials? If TRUE, use_crs must correspond to a projected crs to allow clustering on a positive set of numbers.

pres_x, pres_y

Character. Name of the columns in presences that have the x and y coordinates.

use_crs

Integer. Coordinate reference system (epsg code) to use as the standard for all objects in spatial joins. Using a projected coordinate system removes the sf warning around assuming coordinates are planar if they're not, and can sometimes help with spherical geometry issues. If NULL, the crs of the distribution will be used as the standard (i.e. the presences will be converted to this crs before joining or intersecting).

buf

Integer. Distance in metres to buffer the distribution.

clust_dist

Integer. Distance to base clusters on in clust function if use_clust = TRUE. Needs to be in the units of the projected crs supplied in use_crs, e.g. metres.

overrides

Data frame containing subspecies and poly columns indicating any subspecies to override the overlap checking and force binomials to be updated within the polygon type specified in 'poly', where 'poly' is a vector of values containing any of 'dist', 'mcp' or 'clust' for distribution, minimum convex polygon or cluster respectively.

in_crs

Integer. Coordinate reference system (epsg code) for coordinates in presences.

Value

Data frame equivalent to presences with 'subspecies' updated for taxa where distributions were found and did not overlap with multiple trinomials in presence or distributions, and an additional 'bi_to_tri' column indicating which records were updated.

Details

Will only update binomial occurrence records that fall within a distribution polygon to trinomial if no other trinomial presences or distributions for a species are found intersecting the distribution polygon. This avoids updating binomials to the wrong trinomial where distributions of subspecies overlap. Distribution refers generally to all the polygons used to represent distributions, i.e. distribution files (e.g. from redlist or epbc), MCPs, or cluster polygons.