Calculate a species environment category based on occurrences.

!Experimental! In future this can also be based on distributions.

calc_environ(
  presence,
  land_poly,
  thresh = 95,
  out_file,
  force_new = FALSE,
  pres_x = "long",
  pres_y = "lat",
  in_crs = 4326,
  out_crs = in_crs,
  buf = 0
)

Arguments

presence

Cleaned and filtered dataframe of presences

land_poly

sf polygon object indicating the land in the area of interest.

thresh

Integer. Threshold for percent records to determine if a species is terrestrial, marine or both.

out_file

Character. File path for the results files to be saved.

force_new

Logical. If out_file exists, recreate it?

pres_x, pres_y

Character. Name of the columns in presence that have the x and y coordinates.

in_crs

epsg code for coordinates in presence.

out_crs

epsg code for coordinates in output mcp.

buf

Distance in metres to buffer the land_poly.

Value

data frame saved as rds to out_file path containing environ and land_pc (percent land overlap ) fields for the taxa.