Find file path to relevant taxa distribution (geographic range) parquet files.

For use by filter_by_distribution and fix_spatial_taxonomy functions, and for any other situation needing to return relevant distribution file paths per taxa (e.g. for aoi-taxa and dist in envPIA).

dists_source(
  distrib_dir = fs::path("H:", "data"),
  sources = c("epbc", "expert", "bsa", "seed_centre", "redlist"),
  source_scales = NULL,
  standardise_taxonomy = FALSE,
  target_ranks = c("species", "subspecies"),
  taxonomy = NULL,
  rm_ssp_mismatches = FALSE,
  source_rank = FALSE,
  datatype = "vector"
)

Arguments

distrib_dir

Directory path. Directory containing distribution layers.

sources

Character vector. Distribution sources to use in the distrib_dir, i.e. EPBC, Expert, Birds SA, Redlist, or other.

source_scales

Character vector. Geographic scale of the distribution sources, e.g. state or national. Needs to be the same length as sources, and the position needs to match the position of the source it is referring to. For example, with sources = c("epbc","expert"), the corresponding scale_sources would be c("national","state") where epbc is national and expert is state. Used where different distributions are required at different scales, e.g. state and national distributions both required in filter_by_distribution.

standardise_taxonomy

Logical. Does the taxonomy need to be standardised using a taxonomy object from envClean::make_taxonomy?

target_ranks

Character vector. Target taxonomic ranks for taxonomy standardisation. Either 'species' or 'subspecies', or both.

taxonomy

Taxonomy object returned by envClean::make_taxonomy to use for taxonomy standardisation.

rm_ssp_mismatches

Logical. Remove subspecies mismatches, i.e. where original name was subspecies but was matched to species level by envClean::make_taxonomy. Setting to TRUE is useful where the result is input into filter_by_distribution, as it avoids potentially erroneous filtering of binomial records by trinomial distributions. The downside of setting to TRUE is some relevant distributions may be lost where the match of ssp to spp is correct, e.g. a subspecies is now a new species. Ideally, fixes should be provided for these cases in envClean::make_taxonomy to create the taxonomy object and this parameter set to FALSE.

source_rank

Logical. Rank the distribution sources and filter top ranked? Ranking will be in the order given by sources vector. If !is.null(source_scales) & source_rank == TRUE, then source ranking will done in source_scale groups, i.e. the highest ranked source per scale will be retained. Also, if standardise_taxonomy == TRUE, then ranking will be according to the standardised taxonomy.

datatype

Character. Either 'vector' (default) or 'raster'.

Value

Data frame with relevant distribution sources (ds) and file path (file) per taxa (original_name). If !is.null(source_scales), there will be a file column per scale (e.g. national, state), instead of the single source (ds) column. If standardise_taxonomy == TRUE, the output taxa column will be 'taxa' and not 'original_name'.