Filter species occurrence records by spatial layers representing their distributions at state and or national scales. e.g. IUCN Red List or EPBC (Environmental Protection and Biodiversity Conservation act) distributions.

Filter species occurrence records by spatial layers representing their distributions at state and or national scales. e.g. IUCN Red List or EPBC (Environmental Protection and Biodiversity Conservation act) distributions.

filter_by_distribution(
  presence,
  state,
  national,
  out_file,
  pres_x = "long",
  pres_y = "lat",
  pres_crs = 4326,
  use_crs = NULL,
  buffer_km = 10,
  buffer_override = NULL,
  remove = TRUE,
  state_poly,
  data_ext_poly,
  force_new = FALSE
)

Arguments

presence

Dataframe of presences for a taxa (i.e. containing x,y coordinate columns).

state

Character. File path/s for the relevant state distributions per taxa. If no relevant state distribution, then use NA and filtering will only be conducted using the national distribution. Currently, only parquet files are accepted. Use sfarrow::st_write_parquet to write sf objects to parquet.

national

Character. File path/s for the relevant national distributions per taxa. If no relevant national distribution, then use NA and filtering will only be conducted using the state distribution, but will maintain all interstate records (if any exist). As with state, only parquet files are currently accepted.

out_file

Character. Path for a parquet file to save a pre-filtered dataframe per taxa with records flagged as dist_in = 0 or 1. If previous outputs exist, these outputs will be used to determine what records are new and need to be assessed/filtered, which can save processing time in datasets with many taxa and or complex distributions.

pres_x, pres_y

Character. Name of the columns in presence that have the x and y coordinates.

pres_crs

Integer. Coordinate reference system that the presence coordinates correspond to (epsg code).

use_crs

Integer. Coordinate reference system (epsg code) to use as the standard for all objects in spatial joins or intersections. Using a projected coordinate system removes the sf warning around assuming coordinates are planar if they're not, and can sometimes help with spherical geometry issues. If NULL, the crs of the distribution will be used as the standard (i.e. the presence coords and the data_ext_poly will be converted to this crs before joining or intersecting).

buffer_km

Integer. Distance in kilometres used to buffer the distribution and keep records in that radius. Use 0 for no buffer.

buffer_override

Character vector. Vector of character strings representing distribution sources reflected in file paths for sources that should not be buffered. Used to override the buffer for particular distribution sources where buffering may not be relevant. For example, EPBC distributions already include a 'may occur' area, and if these are stored in a directory called 'epbc_dists', then can use this string to override the buffer for taxa with these distributions.

remove

Logical. Remove out of distribution records? If FALSE, out of distribution records will be flagged (see below).

state_poly

sf object. Polygon representing the state boundary for state and interstate record delineation.

data_ext_poly

sf object. Polygon representing the data extent to use for broad level filtering of taxa with distributions that do not intersect the area of interest.

force_new

Logical. Force new taxa outputs? If TRUE, any previous taxa outputs will be ignored and new filtering/flagging of records conducted.

Value

Presence dataframe with out of range records either removed (if remove == TRUE), or flagged with '1' in the dist_out field (if remove == FALSE). In addition, regardless of remove, a pre-filtered dataframe will be written to the out_file path that contains records flagged as in or out (1 or 0) in the 'dists_in' field.