Make minimum convex polygon (MCP) around occurrence records

Usually used as predict boundary or an Extent of Occurrence (EOO) estimate for a taxa.

make_mcp(
  presence,
  out_file,
  force_new = FALSE,
  pres_x = "long",
  pres_y = "lat",
  in_crs = 4326,
  out_crs = in_crs,
  buf = 0,
  clip = NULL,
  dens_int = NULL
)

Arguments

presence

Cleaned and filtered dataframe of presences

out_file

Character. Path for the mcp to be saved. Will be saved by sfarrow::st_write_parquet(). Currently will not work very well with any full stop in the path. Other file types are changed to .parquet

force_new

Logical. If out_file exists, recreate it?

pres_x, pres_y

Character. Name of the columns in presence that have the x and y coordinates

in_crs

epsg code for coordinates in presence

out_crs

epsg code for coordinates in output mcp. Usually the same as predictors

buf

Distance in metres to buffer the mcp

clip

sf to clip the mcp

dens_int

Numeric. terra::densify() interval argument ("positive number, specifying the desired minimum distance between nodes. The unit is meter for lonlat data, and in the linear unit of the crs for planar data"). Set to NULL to not densify.

Value

sf. out_file saved.