Regional contribution to range (AOO and EOO)

Typically used to determine the contribution of a region to the national range of a species, where a 'region' is a smaller area within the range.

reg_cont(
  taxa,
  presence,
  out_dir,
  mcp_file = NULL,
  grd_file = NULL,
  region_bound,
  force_new = FALSE,
  pres_x = "long",
  pres_y = "lat",
  in_crs = 4326,
  out_crs = in_crs,
  nearest_pop = FALSE,
  clip = NULL
)

Arguments

taxa

Character. Taxa name for use in ConR and outputs.

presence

Cleaned and filtered dataframe of presences.

out_dir

Character. Directory path for the results files to be saved (summary_stats.rds and mcp_noreg.parquet). mcp_noreg will be saved by sfarrow::st_write_parquet(). Currently will not work very well with any full stop in the path. Other file types are changed to .parquet.

mcp_file

Character. Path for the range mcp.

grd_file

Character. Path for the range grid.

region_bound

sf object. Region boundary polygon.

force_new

Logical. If any of the results files in the out_dir exists, recreate it?

pres_x, pres_y

Character. Name of the columns in presence that have the x and y coordinates.

in_crs

epsg code for coordinates in presence.

out_crs

epsg code for coordinates in output 'no region' mcp.

nearest_pop

Logical. Calculate distance to nearest extra-regional population?

clip

sf to clip the 'no region' mcp for comparing the impact of removing regional presences.

Value

summary_stats.rds with aoo and eoo stats, and mcp_noreg.parquet with the mcp excluding regional presences.