Return relevant taxa to a region of interest based on occurrences and or distributions.

Useful for species level analyses seeking to link to relevant trinomials (subspecies, races, varieties, forms, etc), where there may be other irrelevant trinomials occurring outside the region of interest.

reg_taxa(
  bio_df,
  out_dir,
  region_bound,
  force_new = FALSE,
  pres_x = "long",
  pres_y = "lat",
  in_crs = 4326,
  out_crs = in_crs,
  buf = 0,
  taxonomy = targets::tar_read(taxonomy, store = fs::path("H:/dev/out/envCleaned",
    "sa_br_dissolve______0__P50Y__sa_br_dissolve", "90__90__P1Y__subspecies", "clean")),
  listed_df = arrow::read_parquet(fs::path("H:", "data", "taxonomy",
    "all_status_galah.parquet")),
  taxa_ds = dists_source(distrib_dir = data_dir, sources = c("epbc", "expert"),
    source_rank = FALSE, datatype = "vector")
)

Arguments

bio_df

Data frame with taxa column, and lat & long columns.

out_dir

Character. Directory path for the results files to be saved (summary_stats.rds and mcp_noreg.parquet). mcp_noreg will be saved by sfarrow::st_write_parquet(). Currently will not work very well with any full stop in the path. Other file types are changed to .parquet.

region_bound

sf object. Region boundary polygon.

force_new

Logical. If any of the results files in the out_dir exists, recreate it?

pres_x, pres_y

Character. Name of the columns in presence that have the x and y coordinates.

in_crs

epsg code for coordinates in presence.

out_crs

epsg code for coordinates in output grid.

buf

Distance in metres to buffer the region_bound.

taxonomy

Taxonomy object returned by envClean::make_taxonomy in relation to the taxa in bio_df and taxa_ds.

taxa_col

Character. Taxa column in bio_df and taxa_ds (must be the same).

Value

summary_stats.rds with aoo and eoo stats, and mcp_noreg.parquet with the mcp excluding regional presences.