Return relevant taxa to a region of interest based on occurrences and or distributions.

Useful for species level analyses seeking to link to relevant trinomials (subspecies, races, varieties, forms, etc), where there may be other irrelevant trinomials occurring outside the region of interest.

reg_taxa(
  presence,
  distrib_file = NA,
  region_bound,
  taxa,
  pres_x = "long",
  pres_y = "lat",
  pres_crs = 4326,
  use_crs = NULL,
  buf = 0,
  taxonomy,
  remove = FALSE,
  use_mcp = FALSE,
  mcp_file = NULL
)

Arguments

presence

Cleaned and filtered dataframe of presences.

distrib_file

Character. File path/s for the relevant distributions per taxa. If no relevant distribution, then use NA. Currently, only parquet files are accepted. Use sfarrow::st_write_parquet to write sf objects to parquet.

region_bound

sf object. Region boundary polygon.

taxa

Character. Name of the taxa. Needed for joining to taxonomy to obtain species for trinomials (see below).

pres_x, pres_y

Character. Name of the columns in presence that have the x and y coordinates.

use_crs

Integer. Coordinate reference system (epsg code) to use as the standard for all objects in spatial joins or intersections. Using a projected coordinate system removes the sf warning around assuming coordinates are planar if they're not, and can sometimes help with spherical geometry issues. If NULL, the crs of the region_bound will be used as the standard (i.e. the presence coords will be converted to this crs before joining or intersecting).

buf

Distance in metres to buffer the region_bound.

taxonomy

Taxonomy object returned by envClean::make_taxonomy in relation to the taxa. Must have subspecies level taxonomy.

remove

Logical. Remove out of region taxa?

use_mcp

Logical. Use a minimum convex polygon (MCP) around presences to determine region overlap?

mcp_file

Character. If use_mcp == TRUE, optional file path of existing mcp to be used. If use_mcp == TRUE and mcp_file == NULL, a new mcp will be constructed from presence. Currently, only geoparquet files are accepted. Use sfarrow to convert sf objects to geoparquet.

in_crs

epsg code for coordinates in presence.

Value

Data frame with species column, subspecies column, and 'pres_dist', 'distrib_dist' (if relevant) & 'mcp_dist' (if relevant) columns indicating distance in metres from region to closest presence, distribution and mcp respectively. If remove == TRUE, rows with out of region subspecies will not appear.